backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
python-multipart
python-dotenv